# docs/api/openapi.yaml
openapi: "3.1.0"
info:
  title: "MedRecord AI Service API"
  description: |
    API REST para servicio de IA de transcripción y extracción médica en español.

    ## Características
    - Transcripción de audio médico con Whisper (optimizado para español)
    - Extracción de información médica estructurada con GPT-4
    - Validación RAG contra base de conocimiento médico español
    - Sugerencias de códigos CIE-10
    - Detección de interacciones medicamentosas
  version: "1.0.0"
  contact:
    name: "MedRecord AI Team"
    email: "support@medrecord.ai"
servers:
  - url: "http://localhost:8000"
    description: "Desarrollo local"
  - url: "https://api.medrecord.ai"
    description: "Producción"
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/sessions:
    post:
      summary: "Create consultation session"
      description: |
        Create a new session for real-time WebSocket streaming.
        Returns session_id and WebSocket connection URL.
      operationId: createSession
      tags: [Sessions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionCreateRequest"
            example:
              appointment_id: "apt_12345"
              patient_id: "pat_67890"
              doctor_id: "doc_11111"
              specialty: "cardiología"
      responses:
        "200":
          description: "Session created successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    get:
      summary: "List user sessions"
      description: |
        List all sessions for the current user with pagination.
      operationId: listSessions
      tags: [Sessions]
      parameters:
        - name: status_filter
          in: query
          schema:
            type: string
            enum: [active, completed, error]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: "Sessions retrieved"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT token de autenticación"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
        error_code:
          type: string
        timestamp:
          type: string
          format: date-time

    # ═══════════════════════════════════════════════════════════
    # Session Management Schemas
    # ═══════════════════════════════════════════════════════════
    SessionCreateRequest:
      type: object
      properties:
        appointment_id:
          type: string
        patient_id:
          type: string
        doctor_id:
          type: string
        specialty:
          type: string
          default: "medicina general"
        metadata:
          type: object
    SessionStatus:
      type: object
      required: [session_id, status, created_at, updated_at]
      properties:
        session_id:
          type: string
        status:
          type: string
          enum: [active, completed, error, deleted]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        appointment_id:
          type: string
        patient_id:
          type: string
        doctor_id:
          type: string
        specialty:
          type: string
        audio_duration_seconds:
          type: number
        transcription_chunks:
          type: integer
        entities_extracted:
          type: integer
        validations_performed:
          type: integer
        cost_usd:
          type: number
        websocket_url:
          type: string
          description: "WebSocket URL for real-time streaming"
        is_connected:
          type: boolean
    SessionResponse:
      type: object
      properties:
        session:
          $ref: "#/components/schemas/SessionStatus"
        message:
          type: string
    SessionListResponse:
      type: object
      properties:
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/SessionStatus"
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer

    # ═══════════════════════════════════════════════════════════
    # WebSocket Event Schemas
    # ═══════════════════════════════════════════════════════════
...
  responses:
...
    Unauthorized:
      description: "Token inválido o expirado"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...